refactor(plugin-image): clean captionPasteHandler
