+++ packages/editor/src/plugins/text/hooks/use-editable-paste-handler.tsx
@@ -1,6 +1,6 @@
 import { isSelectionWithinList } from '@editor/editor-ui/plugin-toolbar/text-controls/utils/list'
 import { editorPlugins } from '@editor/plugin/helpers/editor-plugins'
-import { onCaptionPasteHandler } from '@editor/plugins/image/utils/on-caption-paste-handler'
+import { captionPasteHandler } from '@editor/plugins/image/utils/caption-paste-handler'
 import { checkIsAllowedNesting } from '@editor/plugins/rows/utils/check-is-allowed-nesting'
 import {
   selectDocument,
@@ -43,7 +43,7 @@ export const useEditablePasteHandler = (args: UseEditablePasteHandlerArgs) => {
       if (!document) return
 
       // special case: pasting in image caption
-      onCaptionPasteHandler({ event, document, files, text, id, dispatch })
+      captionPasteHandler({ event, files, text, id, dispatch })
 
       let media
       // pasting editor document string and insert as plugins
